Assignment 3/requirements.txt
opencv-python==4.8.0.76
numpy==1.24.4
Pillow==10.0.1
tkintertable==1.3.3  # GUI library, part of standard library but good to include
